The features page was annoyingly slow.

Instead of loading ~20 YouTube videos, now using jQuery to load them in a popup when you click them. 

It took me a day but the results:
* from 9.5 sec, 2.6mb to under 0.9 sec, 2mb in chromium
* from 20 sec, 27mb (!) to under 4 sec, 2mb in firefox

That's locally, we will have to see how it does on staging.

It looks differently, of course... But not too bad.