@jospoortvliet 2016-04-20
tested by @danimo too. Going for staging.