@phil-davis 2016-04-20
Very good. I have users in Nepal on 256, 384, 512Kbps connections. It is REALLY annoying when pages auto-download heaps of content when the user just wants to browse around and then click the content that they want to look at. And of course people with tablet/phones on mobile plans i an country who do not appreciate losing 20MB of their quota to browse to a page.