@jospoortvliet 2016-04-21
@phil-davis yeah, this will hopefully make it more feasible to check out our site in places like that. We're going to move to a new server too, supporting http2 and other stuff, should speed up things as well. And I'll look over some other pages to speed those up, too - most are fine but the YouTube stuff is big and heavy. Then again, I think pages with a single video are fine...